feat: implement modular kanban components, unified API client, and state management for the dashboard shell
