+++ apps/web/src/lib/boards.ts
@@ -1,4 +1,4 @@
-import axios, { AxiosError } from "axios";
+import { apiRequest } from "./api-client";
 
 import type {
   BoardActivityItem,
@@ -9,53 +9,6 @@ import type {
   BoardSummary,
   BoardVisibility,
 } from "@/components/dashboard/board-types";
-
-type ApiErrorResponse = {
-  message?: string | string[];
-};
-
-function getApiBaseUrl() {
-  return process.env.NEXT_PUBLIC_API_URL ?? "http://localhost:3001";
-}
-
-function normalizeApiErrorMessage(message: ApiErrorResponse["message"]) {
-  if (Array.isArray(message)) {
-    return message[0] ?? "Something went wrong. Please try again.";
-  }
-
-  return message ?? "Something went wrong. Please try again.";
-}
-
-const apiClient = axios.create({
-  baseURL: getApiBaseUrl(),
-  withCredentials: true,
-  headers: {
-    "Content-Type": "application/json",
-  },
-});
-
-async function apiRequest<T>(
-  path: string,
-  init: {
-    body?: unknown;
-    method: "GET" | "POST" | "PATCH" | "DELETE";
-  },
-): Promise<T> {
-  try {
-    const response = await apiClient.request<T>({
-      url: path,
-      method: init.method,
-      data: init.body,
-    });
-
-    return response.data;
-  } catch (error) {
-    const axiosError = error as AxiosError<ApiErrorResponse>;
-    throw new Error(
-      normalizeApiErrorMessage(axiosError.response?.data?.message),
-    );
-  }
-}
 
 export function listWorkspaceBoards(workspaceId: string) {
   return apiRequest<BoardSummary[]>(`/workspaces/${workspaceId}/boards`, {
