+++ apps/web/src/lib/workspaces.ts
@@ -1,4 +1,3 @@
-import axios, { AxiosError } from "axios";
 import type {
   WorkspaceActivityItem,
   WorkspaceDetail,
@@ -7,58 +6,12 @@ import type {
   WorkspaceRole,
   WorkspaceSummary,
 } from "@/components/dashboard/workspace-types";
-
-type ApiErrorResponse = {
-  message?: string | string[];
-};
+import { apiRequest } from "./api-client";
 
 export type WorkspaceSummaryResponse = WorkspaceSummary;
 export type WorkspaceDetailResponse = WorkspaceDetail;
 export type WorkspaceMemberResponse = WorkspaceMember;
 export type WorkspaceInviteResponseBody = WorkspaceInviteResponse;
-
-function getApiBaseUrl() {
-  return process.env.NEXT_PUBLIC_API_URL ?? "http://localhost:3001";
-}
-
-function normalizeApiErrorMessage(message: ApiErrorResponse["message"]) {
-  if (Array.isArray(message)) {
-    return message[0] ?? "Something went wrong. Please try again.";
-  }
-
-  return message ?? "Something went wrong. Please try again.";
-}
-
-const apiClient = axios.create({
-  baseURL: getApiBaseUrl(),
-  withCredentials: true,
-  headers: {
-    "Content-Type": "application/json",
-  },
-});
-
-async function apiRequest<T>(
-  path: string,
-  init: {
-    body?: unknown;
-    method: "GET" | "POST" | "PATCH" | "DELETE";
-  },
-): Promise<T> {
-  try {
-    const response = await apiClient.request<T>({
-      url: path,
-      method: init.method,
-      data: init.body,
-    });
-
-    return response.data;
-  } catch (error) {
-    const axiosError = error as AxiosError<ApiErrorResponse>;
-    throw new Error(
-      normalizeApiErrorMessage(axiosError.response?.data?.message),
-    );
-  }
-}
 
 export function createWorkspace(input: { name: string }) {
   return apiRequest<WorkspaceSummaryResponse>("/workspaces", {
